+++ packages/env.d.ts
@@ -1,5 +1,6 @@
-/// <reference types="vite/client" />
-
 interface ImportMeta {
-  readonly env: ImportMetaEnv;
+  env: {
+    PROD: boolean;
+    DEV: boolean;
+  };
 }