+++ packages/peer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @starbeam/peer
 
+## 1.0.6
+
+### Patch Changes
+
+- Fix ESM package mistakes
+
 ## 1.0.5
 
 ### Patch Changes