+++ packages/timeline/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @starbeam/timeline
 
+## 0.5.6
+
+### Patch Changes
+
+- Fix ESM package mistakes
+- Updated dependencies
+  - @starbeam/debug@0.5.6
+  - @starbeam/peer@1.0.6
+  - @starbeam/verify@0.5.6
+
 ## 0.5.5
 
 ### Patch Changes